requirements.txt
Django==2.1.11
django-extensions==2.2.1
django-js-asset==1.2.2
django-mptt==0.10.0
django-pyodbc-azure==2.1.0.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pyodbc==4.0.27
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
